# api/docs/swagger.yaml
definitions:
  ChangeEmailQuery:
    properties:
      email:
        description: New email address
        example: newmail@newcomp.com
        type: string
    required:
    - email
    type: object
  ChangePwdQuery:
    properties:
      new_password:
        description: New password
        example: new_password
        type: string
      old_password:
        description: Old password
        example: old_password
        type: string
    required:
    - new_password
    - old_password
    type: object
  ConfirmEmailChangeQuery:
    properties:
      token:
        description: Change token
        example: my_change_token
        type: string
    required:
    - token
    type: object
...
  JSendError:
    properties:
      message:
        description: Error message
        example: Internal server error
        type: string
      status:
        description: Status
        example: error
        type: string
    type: object
  JSendFailure-ErrorResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/ErrorResponse'
        description: Data with error message(s)
      status:
        description: Status 'fail'
        example: fail
        type: string
    type: object
  JSendFailure-ValidationResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/ValidationResponse'
        description: Data with error message(s)
      status:
        description: Status 'fail'
        example: fail
        type: string
    type: object
  JSendSuccess-LoginResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/LoginResponse'
        description: Data with success message(s)
      status:
        description: Status 'success'
        example: success
        type: string
    type: object
...
  JSendSuccess-map_string_string:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/map_string_string'
        description: Data with success message(s)
      status:
        description: Status 'success'
        example: success
        type: string
    type: object
  LoginQuery:
    properties:
      login:
        example: joe@me.com
        type: string
      password:
        example: password
        type: string
      role:
        enum:
        - admin
        - user
        - approver
        example: user
        type: string
    required:
    - login
    - password
    type: object
...
  ResetPwdQuery:
    properties:
      email:
        description: Email address
        example: joe@me.com
        type: string
    required:
    - email
    type: object
...
info:
  contact: {}
paths:
...
  /user:
    delete:
      description: |-
        The account will be soft deleted.
        If the user is the last admin, the account cannot be deleted.
        If a user is soft-deleted, the account will be permanently deleted in the future.
      produces:
      - application/json
      responses:
        "200":
          description: Password reset
          schema:
            $ref: '#/definitions/JSendSuccess-map_string_string'
        "400":
          description: Last admin account
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/JSendError'
      security:
      - Bearer: []
      summary: Delete user account
      tags:
      - User
  /user/email:
    patch:
      description: Requests an email change for the user. An email change token will
        be sent to the new email address.
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ChangeEmailQuery'
      produces:
      - application/json
      responses:
        "200":
          description: Email change request token sent
          schema:
            $ref: '#/definitions/JSendSuccess-map_string_string'
        "400":
          description: Bad request/invalid email/already in use
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "422":
          description: Bad query format
          schema:
            $ref: '#/definitions/JSendFailure-ValidationResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/JSendError'
      security:
      - Bearer: []
      summary: Request email change for the user
      tags:
      - User
  /user/email/confirm:
    post:
      description: |-
        Confirms an email change for the user.
        The new email address will be active on the next login.
        You have to login (authenticate) with the old email address to confirm the change.
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ConfirmEmailChangeQuery'
      produces:
      - application/json
      responses:
        "200":
          description: Email changed
          schema:
            $ref: '#/definitions/JSendSuccess-map_string_string'
        "400":
          description: Invalid token
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "403":
          description: Token expired
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "404":
          description: No email change request found
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "422":
          description: Bad query format
          schema:
            $ref: '#/definitions/JSendFailure-ValidationResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/JSendError'
      security:
      - Bearer: []
      summary: Confirm email change for the user
      tags:
      - User
  /user/login:
    post:
      description: |-
        Acciqures a JWT token for the user to access the API
        Only active users can login
        Users can downgrade their role by providing the role in the request (optional)
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/LoginQuery'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            $ref: '#/definitions/JSendSuccess-LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "403":
          description: User is not active
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "422":
          description: Bad query format
          schema:
            $ref: '#/definitions/JSendFailure-ValidationResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/JSendError'
      summary: Login for the API to get JWT token
      tags:
      - Login
  /user/password:
    patch:
      description: |-
        Changes the password for the user. The old password must be provided.
        The new password will be active on the next login.
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ChangePwdQuery'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed
          schema:
            $ref: '#/definitions/JSendSuccess-map_string_string'
        "400":
          description: Wrong old password/invalid new password
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "422":
          description: Bad query format
          schema:
            $ref: '#/definitions/JSendFailure-ValidationResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/JSendError'
      security:
      - Bearer: []
      summary: Change password for the user
      tags:
      - User
  /user/password/init:
    post:
      description: |-
        Confirms a password reset or first password set for the user.
        The API will always return the same message (400) on auth errors to prevent email enumeration.
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ResetConfirmPwdQuery'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset
          schema:
            $ref: '#/definitions/JSendSuccess-map_string_string'
        "400":
          description: Bad request/invalid token
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "403":
          description: Token expired
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "422":
          description: Bad query format
          schema:
            $ref: '#/definitions/JSendFailure-ValidationResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/JSendError'
      summary: Confirm password reset or first password set
      tags:
      - Login
  /user/password/reset:
    post:
      description: |-
        Requests a password reset for the user. A password reset token will be sent to the user's email.
        Password reset tokens are valid for a limited time.
        The API will always return the same message (200) to prevent email enumeration.
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ResetPwdQuery'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset token sent
          schema:
            $ref: '#/definitions/JSendSuccess-map_string_string'
        "422":
          description: Bad query format
          schema:
            $ref: '#/definitions/JSendFailure-ValidationResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/JSendError'
      summary: Request password reset
      tags:
      - Login
  /user/password/reset/confirm:
    post:
      description: |-
        Confirms a password reset or first password set for the user.
        The API will always return the same message (400) on auth errors to prevent email enumeration.
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ResetConfirmPwdQuery'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset
          schema:
            $ref: '#/definitions/JSendSuccess-map_string_string'
        "400":
          description: Bad request/invalid token
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "403":
          description: Token expired
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "422":
          description: Bad query format
          schema:
            $ref: '#/definitions/JSendFailure-ValidationResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/JSendError'
      summary: Confirm password reset or first password set
      tags:
      - Login
...
  /user/refresh-token:
    post:
      description: Refreshes the JWT token for the user to access the API
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/RefreshQuery'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            $ref: '#/definitions/JSendSuccess-LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "403":
          description: Not active/role invalid/user deleted
          schema:
            $ref: '#/definitions/JSendFailure-ErrorResponse'
        "422":
          description: Bad query format
          schema:
            $ref: '#/definitions/JSendFailure-ValidationResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/JSendError'
      summary: Refresh JWT token
      tags:
      - Login
swagger: "2.0"